move PaymentStatus component from OrderList dir - it is global component
